generate statement removed now element generating the statement
